IFC-751 save diff in smaller pieces to prevent OOM
